+++ Changelog.md
@@ -5,6 +5,8 @@ Changelog
 
 * Added support for ```scripts``` and ```styles``` properties in the component.json file. As used in node
 * Throw an exception instead of deleting the cache when the dependency mapping returned from bower is invalid
+* Added documentation on how to process images/fonts in css files
+* Added new option `assetic.nest_dependencies` to configure the behavior of nested dependencies in the generated assetic resources
 
 ### 0.6 (2013-04-06)